+++ packages/middleware/CHANGELOG.md
@@ -1,4 +1,17 @@
-# @vue-storefront/middleware
+# Change log
+
+## 3.6.1
+
+### Patch Changes
+
+- 3335b9b48: `getApiClient` helper returns now ApiClient interface
+
+  Usage:
+
+  ```typescript
+  const sapcc = context.getApiClient<Api, Config, Client>("sapcc");
+  // typeof sapcc === ApiClient<Api, Config, Client>
+  ```
 
 ## 3.6.0